feat: add Git guide content
Motivation:
It might be useful to get a common ground when it comes to how Git work. So a documentation on how things work might be needed.

Add a documentatin on common steps in Git.

Add a cheat sheet summarising the common steps in Git.

Add a description on how Conventional Commits work for better commit messages.

Add a description of what Git is and what it is used for.